Add focal loss option and improve configuration in Attention-UNET
This commit provides the option to use focal loss during training in the Attention-UNET model. The decision to use focal loss can be specified in the configuration file. The code now adjusts according to the selected loss method in both training and validation operations. Assistance for the AttentionBlock in model loading was also added. The setup file was created, import statements were refined, and custom validation IOU method was slightly renamed.
Also added setup.py for pip install.